backend/requirements.txt
boto3==1.23.1
botocore==1.26.1
click==8.1.3
Flask==2.1.2
Flask-Cors==3.0.10
google-cloud-bigtable==2.10.0
google-cloud-core==2.3.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
python-dateutil==2.8.2
s3transfer==0.5.2
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
